partially break up conftest
