+++ tests/actions/commands/test_close.py
@@ -1,5 +1,5 @@
 from cibo.models import ClientLoginState, Message, MessageRoute
-from tests.conftest import CloseActionFactory
+from tests.actions.conftest import CloseActionFactory
 
 
 class TestCloseAction(CloseActionFactory):
