+++ tests/actions/commands/test_inventory.py
@@ -1,6 +1,6 @@
 from cibo.models import ClientLoginState, Message, MessageRoute
 from cibo.models.data import Player
-from tests.conftest import InventoryActionFactory
+from tests.actions.conftest import InventoryActionFactory
 
 
 class TestInventoryAction(InventoryActionFactory):
