+++ tests/actions/test_error.py
@@ -1,5 +1,5 @@
 from cibo.models import Message, MessageRoute
-from tests.conftest import ErrorActionFactory
+from tests.actions.conftest import ErrorActionFactory
 
 
 class TestErrorAction(ErrorActionFactory):
